+++ src/App.js
@@ -1,8 +1,8 @@
-import './App.css';
+import './css/App.css';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import Society from './pages/society';
 
-import FlipCard from './components/FlipCard';
+import CardsList from './components/CardsList';
 import Slider from './components/Carousel/index.js';
 
 import Footer from './components/Global/Footer';
@@ -15,7 +15,7 @@ function App() {
         <Route exact path='/' element={<div className="App">
           <Header />
           <Slider />
-          <FlipCard />
+          <CardsList />
           <Footer />
         </div>}
         />
